Added default tsconfig.json to react application template to make react typescript work better out of the box.
Added logic around conditionally adding tailwind css to that is is not
included if the profiler is set to use css

+++ src/index.ts
@@ -97,21 +97,22 @@ export const buildProject = async (project: Project) => {
           name
         )
 
-        fs.unlinkSync(path.normalize(`${name}/src/index.css`))
+        if (profiler.CSS_EXTENSION === 'scss') {
+          fs.unlinkSync(path.normalize(`${name}/src/index.css`))
+          await ncp(
+              path.join(__dirname, '../templates/application-extras/tailwind'),
+              name
+          )
 
-        await ncp(
-          path.join(__dirname, '../templates/application-extras/tailwind'),
-          name
-        )
-
-        const packageJSON = JSON.parse(
-          fs.readFileSync(path.join(name, 'package.json'), 'utf8')
-        )
-        packageJSON.devDependencies.tailwindcss = '^2.0.2'
-        fs.writeFileSync(
-          path.join(name, 'package.json'),
-          JSON.stringify(packageJSON, null, 2)
-        )
+          const packageJSON = JSON.parse(
+              fs.readFileSync(path.join(name, 'package.json'), 'utf8')
+          )
+          packageJSON.devDependencies.tailwindcss = '^2.0.2'
+          fs.writeFileSync(
+              path.join(name, 'package.json'),
+              JSON.stringify(packageJSON, null, 2)
+          )
+        }
       }
       break
   }
